Improve documentation for conditional compilation
Having read uglifyjs' documentation, I came to this repo looking for how to use the equivalent of the `--define` command line option. I almost was unable to find how to do so, but #12 saved me, indicating that the documentation was just not great on how to use this feature.

So this change would hopefully just help people like me who searched the page for `--define` and didn't know what to do when they couldn't find it.