handle minification failure gracefully
Fixes: https://github.com/gruntjs/grunt-contrib-uglify/issues/409

Whenever the minification of a file fails, the task fails with the
error 'min' is not available in 'undefined'.

This patch allows the task to continue to the next file without
making any changes to the contents of the file in the destination.
Also, this introduces a new configuration option `skipErrors`, so
that the users can specify that in the task level options to override
the default behaviour.